+++ c/gemm-dense-openmp64.c
@@ -7,20 +7,20 @@
 #include <omp.h>
 #endif
 
-static void fill_random(float *A, const int64_t n, const int64_t m) {
+static void fill_random(double *A, const int64_t n, const int64_t m) {
 
   int64_t i, j;
   for (i = 0; i < n; ++i) {
     for (j = 0; j < m; ++j) {
-      A[i * m + j] = (float)rand() / (float)RAND_MAX;
+      A[i * m + j] = (double)rand() / (double)RAND_MAX;
     }
   }
 }
 
-static void gemm(float *A, float *B, float *C, const int64_t A_rows,
+static void gemm(double *A, double *B, double *C, const int64_t A_rows,
                  const int64_t A_cols, const int64_t B_cols) {
   int64_t i, k, j;
-  float temp;
+  double temp;
 #ifdef _OPENMP
 #pragma omp parallel for default(shared) private(i, k, j, temp)
 #endif
@@ -37,21 +37,22 @@ static void gemm(float *A, float *B, float *C, const int64_t A_rows,
 static struct timespec print_dtime(struct timespec start, const char *process) {
   struct timespec end;
   clock_gettime(CLOCK_MONOTONIC_RAW, &end);
-  const float dtime = ((float)((end.tv_sec - start.tv_sec) * 1000000 +
-                               (end.tv_nsec - start.tv_nsec) / 1000)) /
-                      1E6;
+  const double dtime = ((double)((end.tv_sec - start.tv_sec) * 1000000 +
+                                 (end.tv_nsec - start.tv_nsec) / 1000)) /
+                       1E6;
 
   printf("Time to %s = %f s\n", process, dtime);
   return end;
 }
 
-static float dtime(struct timespec start, struct timespec end) {
-  return ((float)((end.tv_sec - start.tv_sec) * 1000000 +
-                  (end.tv_nsec - start.tv_nsec) / 1000)) /
+static double dtime(struct timespec start, struct timespec end) {
+  return ((double)((end.tv_sec - start.tv_sec) * 1000000 +
+                   (end.tv_nsec - start.tv_nsec) / 1000)) /
          1E6;
 }
 
-static void print_matrix(float *A, const int64_t A_rows, const int64_t A_cols) {
+static void print_matrix(double *A, const int64_t A_rows,
+                         const int64_t A_cols) {
 
   int64_t i, j;
   printf("[");
@@ -93,14 +94,16 @@ int main(int argc, char *argv[]) {
   struct timespec start, tmp;
   clock_gettime(CLOCK_MONOTONIC_RAW, &start);
 
-  float *A = (float *)malloc((size_t)A_rows * (size_t)A_cols * sizeof(float));
+  double *A =
+      (double *)malloc((size_t)A_rows * (size_t)A_cols * sizeof(double));
   tmp = print_dtime(start, "allocate A");
 
-  float *B = (float *)malloc((size_t)B_rows * (size_t)B_cols * sizeof(float));
+  double *B =
+      (double *)malloc((size_t)B_rows * (size_t)B_cols * sizeof(double));
   tmp = print_dtime(tmp, "allocate B");
 
   // value-init to zero
-  float *C = (float *)calloc((size_t)A_rows * (size_t)B_cols, sizeof(float));
+  double *C = (double *)calloc((size_t)A_rows * (size_t)B_cols, sizeof(double));
   tmp = print_dtime(tmp, "initialize C");
 
   fill_random(A, A_rows, A_cols);
@@ -115,7 +118,7 @@ int main(int argc, char *argv[]) {
 
   if (steps > 1) {
 
-    float average_time = 0;
+    double average_time = 0;
     struct timespec start_i, end_i;
 
     for (i = 1; i < steps; ++i) {
